finally add the youtube image cache to the article uploads

+++ public_html/includes/ajax/delete_image.php
@@ -5,18 +5,25 @@ define("APP_ROOT", dirname ( dirname ( dirname(__FILE__) ) ) );
 
 require APP_ROOT . "/includes/bootstrap.php";
 
+// only logged in accounts can do this too, it's only for articles
+if (!isset($_SESSION['user_id']) || $_SESSION['user_id'] == 0)
+{
+	die('You shouldn\'t be here. You need to be logged in.');
+}
+
 if (isset($_POST['image_id']) && is_numeric($_POST['image_id']))
 {
-	// get the image 
-	$qry1 = "SELECT `id`, `filename`, `filetype` FROM `article_images` WHERE `id` = ?";
+	$qry1 = "SELECT `id`, `filename`, `filetype`,`youtube_cache`, `uploader_id` FROM `article_images` WHERE `id` = ?";
 	$grabber = $dbl->run($qry1, array($_POST['image_id']))->fetch();
 
 	$qry2 = "DELETE FROM `article_images` WHERE `id` = ?";
 	$result = $dbl->run($qry2, array($_POST['image_id']));
 	if(isset($result))
 	{
-		if(unlink(APP_ROOT . '/uploads/articles/article_media/' . $grabber['filename']))
+		if ($grabber['youtube_cache'] == 0)
 		{
+			unlink(APP_ROOT . '/uploads/articles/article_media/' . $grabber['filename']);
+
 			if ($grabber['filetype'] == 'gif')
 			{
 				$static_filename = str_replace('.gif', '_static.jpg', $grabber['filename']);
@@ -26,9 +33,16 @@ if (isset($_POST['image_id']) && is_numeric($_POST['image_id']))
 			{
 				unlink(APP_ROOT . '/uploads/articles/article_media/thumbs/' . $grabber['filename']);
 			}
-			unset($_SESSION['uploads'][$grabber['id']]);
-			echo "YES";
 		}
+		else
+		{
+			if (file_exists(APP_ROOT . '/cache/youtube_thumbs/' . $grabber['filename']))
+			{
+				unlink(APP_ROOT . '/cache/youtube_thumbs/' . $grabber['filename']);
+			}				
+		}
+		unset($_SESSION['uploads'][$grabber['id']]);
+		echo "YES";
 	}
 
 	else
